:recycle: move PurchaseResults and SellResults to transactions file
