+++ host/nation/types/transactions.py
@@ -1,0 +1,11 @@
+from enum import Enum, auto
+
+
+class PurchaseResult(Enum):
+    SUCCESS = auto()
+    INSUFFICIENT_FUNDS = auto()
+
+
+class SellResult(Enum):
+    SUCCESS = auto()
+    INSUFFICIENT_AMOUNT = auto()
